pyproject.toml
[project]
name = "opsboard"
version = "0.1.0"
description = "OpsBoard MVP for Able to compete challenge"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
  "fastapi>=0.116.0",
  "jinja2>=3.1.4",
  "python-multipart>=0.0.9",
  "supabase>=2.8.1",
  "uvicorn>=0.30.6",
  "httpx>=0.28.1",
  "pydantic>=2.9.2",
  "google-genai>=1.0.0",
]

[project.optional-dependencies]
dev = [
  "pytest>=8.3.3",
  "pytest-bdd>=8.1.0",
  "ruff>=0.7.1",
]

[tool.pytest.ini_options]
pythonpath = ["."]
testpaths = ["tests"]

[tool.ruff]
line-length = 100
target-version = "py311"
exclude = [".agents", ".snapshots"]

[build-system]
requires = ["setuptools>=69", "wheel"]
build-backend = "setuptools.build_meta"

[tool.setuptools.packages.find]
where = ["."]
include = ["app*"]
